refactor to separate workflow, testing, InMemory, Zoom, and API
